Enhance NFT query generation for price and trait searches
- Improve system prompt to handle price-related queries correctly
- Add explicit instructions for case-insensitive trait matching
- Include examples for proper query structure with price ranges and sorting
- Update price field handling to use 'exists' clause for listed items
- Refine trait search to use separate nested queries for multiple conditions
- Add safeguards against potential security vulnerabilities in query generation

+++ amplify/backend.ts
@@ -141,7 +141,7 @@ const indexMapping = {
         type: "text",
         fields: { keyword: { type: "keyword", ignore_above: 256 } },
       },
-      price: { type: "long" },
+      price: { type: "float" },
       traits: {
         type: "nested",
         properties: {
